requirements.txt
streamlit
langchain
chromadb
numpy
langchain-huggingface
langchain-google-genai 
google-generativeai
langchain-community
pypdf
sentence-transformers
huggingface_hub